+++ werkzeug/routing.py
@@ -107,7 +107,7 @@ from functools import partial
 from pprint import pformat
 from threading import Lock
 
-from werkzeug.urls import url_encode, url_quote, url_join, fast_url_quote
+from werkzeug.urls import url_encode, url_quote, url_join, _fast_url_quote
 from werkzeug.utils import redirect, format_string
 from werkzeug.exceptions import HTTPException, NotFound, MethodNotAllowed, \
      BadHost
@@ -1230,7 +1230,7 @@ class BaseConverter(object):
         return value
 
     def to_url(self, value):
-        return fast_url_quote(text_type(value).encode(self.map.charset))
+        return _fast_url_quote(text_type(value).encode(self.map.charset))
 
 
 class UnicodeConverter(BaseConverter):